+++ src/api/users/profile-model.ts
@@ -1,15 +1,6 @@
 import db from '../../data/dbConfig';
 import { timestamp } from '../common/util';
-export interface ProfileType {
-    id: number;
-    user_id: number;
-    name: string;
-    avatar: string;
-    email: string;
-    dob: Date;
-    config: JSON;
-    updated_at: any;
-}
+import { ProfileType } from '../common/types'
 
 export default {
     findBy,
@@ -22,27 +13,27 @@ export default {
     yourProfile
 }
 
-function byId(id:number):Promise<ProfileType>  {
+function byId(id: number): Promise<ProfileType> {
     return db("profiles")
-    .where({ id })
-    .first();
+        .where({ id })
+        .first();
 }
 
-async function findBy(filter:Partial<ProfileType>):Promise<ProfileType[]> {
+async function findBy(filter: Partial<ProfileType>): Promise<ProfileType[]> {
     return await db("profiles").where(filter);
 }
 
-async function findByUsername(username:string):Promise<ProfileType> {
+async function findByUsername(username: string): Promise<ProfileType> {
     return await db("profiles")
-    .join("users", "users.id", "=", "profiles.user_id")
-    .where({ username })
-    .first()
-    .then((profile:any) => {
-        // remove hashed password
-        if (!profile) return {}
-        let { password, ...rest } = profile;
-        return rest;
-    });
+        .join("users", "users.id", "=", "profiles.user_id")
+        .where({ username })
+        .first()
+        .then((profile: any) => {
+            // remove hashed password
+            if (!profile) return {}
+            let { password, ...rest } = profile;
+            return rest;
+        });
 }
 
 async function findByUserId(user_id: number): Promise<any> {
@@ -59,38 +50,38 @@ async function findByUserId(user_id: number): Promise<any> {
 }
 
 
-async function addByUsername(username:string):Promise<ProfileType|undefined>  {
+async function addByUsername(username: string): Promise<ProfileType | undefined> {
     const user_id = (await db("users").where({ username })?.first())?.id;
     if (!user_id) {
         return undefined
     };
     return db("profiles")
-    .insert({ user_id })
-    .then(() => findByUsername(username));
+        .insert({ user_id })
+        .then(() => findByUsername(username));
 }
 
-function add(profile:Partial<ProfileType>):Promise<ProfileType> {
+function add(profile: Partial<ProfileType>): Promise<ProfileType> {
     return db("profiles")
-    .insert(profile)
-    .then((ids:number[]) =>{
-        const [id] = ids;
-        return byId(id);
-    });
+        .insert(profile)
+        .then((ids: number[]) => {
+            const [id] = ids;
+            return byId(id);
+        });
 }
 
-function update(id:number, data:Partial<ProfileType>) {
-    return db("profiles").where({ id }).update({...data, updated_at: timestamp() });
+function update(id: number, data: Partial<ProfileType>) {
+    return db("profiles").where({ id }).update({ ...data, updated_at: timestamp() });
 }
 
 
-async function yourProfile(id?: number, username?:string) {
-    let params:any = {};
+async function yourProfile(id?: number, username?: string) {
+    let params: any = {};
     if (!id || !username) return;
     if (id) params.user_id = id;
     if (username) params.username = username;
     return await db("profiles")
-    .join("users", "users.id", "=", "profiles.user_id")
-    .select("username", "name", "avatar", "email", "dob", "created_at")
-    .where(params)
-    .first();
+        .join("users", "users.id", "=", "profiles.user_id")
+        .select("username", "name", "avatar", "email", "dob", "created_at")
+        .where(params)
+        .first();
 }
